# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IQmeetEQ Workshop Companion API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface LoginRequest {
  email: string;
}

export interface Participant {
  id: number;
  email: string;
  createdAt: string;
  lastLoginAt: string;
}

export interface LoginResponse {
  participant: Participant;
  isNew: boolean;
}

export type SectionWithStatusType =
  (typeof SectionWithStatusType)[keyof typeof SectionWithStatusType];

export const SectionWithStatusType = {
  exercise: "exercise",
  reference: "reference",
  locked: "locked",
} as const;

export interface SectionWithStatus {
  id: string;
  title: string;
  description: string;
  day: number;
  order: number;
  tier: number;
  type: SectionWithStatusType;
  unlocked: boolean;
  hasCode: boolean;
}

export interface UnlockRequest {
  code: string;
}

export interface UnlockResponse {
  sectionId: string;
  sectionTitle: string;
  alreadyUnlocked: boolean;
}

export interface Note {
  id: number;
  sectionId: string;
  fieldKey: string;
  content: string;
  updatedAt: string;
}

export interface SaveNoteRequest {
  fieldKey: string;
  content: string;
}

export interface WorkflowMap {
  id: number;
  name: string;
  frequency: string;
  currentSteps: string[];
  redesignedSteps: string[];
  verificationCheckpoints: string;
  stopConditions: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateWorkflowMapRequest {
  name: string;
  frequency?: string;
  currentSteps?: string[];
  redesignedSteps?: string[];
  verificationCheckpoints?: string;
  stopConditions?: string;
}

export interface UpdateWorkflowMapRequest {
  name?: string;
  frequency?: string;
  currentSteps?: string[];
  redesignedSteps?: string[];
  verificationCheckpoints?: string;
  stopConditions?: string;
}

export interface Feedback {
  id: number;
  content: string;
  updatedAt: string;
}

export interface SaveFeedbackRequest {
  content: string;
}

export interface AdminLoginRequest {
  password: string;
}

export interface ParticipantEngagement {
  id: number;
  email: string;
  lastLoginAt: string;
  createdAt: string;
  sectionsOpened: number;
  hasNotes: boolean;
  hasWorkflowMaps: boolean;
  hasFeedback: boolean;
}

export interface AdminSection {
  id: string;
  title: string;
  code: string;
  codeActive: boolean;
  cohort?: string;
  tier: number;
}

export interface UpdateSectionCodeRequest {
  sectionId: string;
  code: string;
  codeActive: boolean;
  cohort?: string;
}

export interface SafariWorksheet {
  id: number;
  name: string;
  sortOrder: number;
  active: boolean;
}

export interface CreateSafariWorksheetRequest {
  name: string;
  sortOrder?: number;
}

export interface UpdateSafariWorksheetRequest {
  name?: string;
  sortOrder?: number;
  active?: boolean;
}

export interface AdminFeedback {
  id: number;
  participantEmail: string;
  content: string;
  updatedAt: string;
}

export interface AdminStats {
  totalParticipants: number;
  activeToday: number;
  sectionsUnlocked: number;
  totalWorkflowMaps: number;
  feedbackSubmissions: number;
}