# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: IQmeetEQ Workshop Companion API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login or register participant by email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Invalid email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current participant
      responses:
        "200":
          description: Current participant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Participant"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /notes/{sectionId}:
    get:
      operationId: getNotesBySection
      tags: [notes]
      summary: Get notes for a specific section
      parameters:
        - name: sectionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notes for section
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"

    put:
      operationId: saveNote
      tags: [notes]
      summary: Save a note for a section
      parameters:
        - name: sectionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveNoteRequest"
      responses:
        "200":
          description: Note saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginRequest"
      responses:
        "200":
          description: Admin login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/sections:
    get:
      operationId: getAdminSections
      tags: [admin]
      summary: Get all sections with their codes
      responses:
        "200":
          description: Sections with codes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminSection"

    put:
      operationId: updateSectionCode
      tags: [admin]
      summary: Update a section's code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSectionCodeRequest"
      responses:
        "200":
          description: Code updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSection"
  /admin/safari-worksheets:
    get:
      operationId: getSafariWorksheets
      tags: [admin]
      summary: Get all safari worksheets
      responses:
        "200":
          description: Worksheets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SafariWorksheet"

    post:
      operationId: createSafariWorksheet
      tags: [admin]
      summary: Create safari worksheet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSafariWorksheetRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SafariWorksheet"
  /admin/safari-worksheets/{id}:
    put:
      operationId: updateSafariWorksheet
      tags: [admin]
      summary: Update safari worksheet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSafariWorksheetRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SafariWorksheet"

    delete:
      operationId: deleteSafariWorksheet
      tags: [admin]
      summary: Delete safari worksheet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    LoginRequest:
      type: object
      properties:
        email:
          type: string
      required:
        - email
    LoginResponse:
      type: object
      properties:
        participant:
          $ref: "#/components/schemas/Participant"
        isNew:
          type: boolean
      required:
        - participant
        - isNew
    Participant:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - createdAt
        - lastLoginAt
...
    Note:
      type: object
      properties:
        id:
          type: integer
        sectionId:
          type: string
        fieldKey:
          type: string
        content:
          type: string
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - sectionId
        - fieldKey
        - content
        - updatedAt
    SaveNoteRequest:
      type: object
      properties:
        fieldKey:
          type: string
        content:
          type: string
      required:
        - fieldKey
        - content
...
    AdminLoginRequest:
      type: object
      properties:
        password:
          type: string
      required:
        - password
...
    AdminSection:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        code:
          type: string
        codeActive:
          type: boolean
        cohort:
          type: string
        tier:
          type: integer
      required:
        - id
        - title
        - code
        - codeActive
        - tier
    UpdateSectionCodeRequest:
      type: object
      properties:
        sectionId:
          type: string
        code:
          type: string
        codeActive:
          type: boolean
        cohort:
          type: string
      required:
        - sectionId
        - code
        - codeActive
    SafariWorksheet:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sortOrder:
          type: integer
        active:
          type: boolean
      required:
        - id
        - name
        - sortOrder
        - active
    CreateSafariWorksheetRequest:
      type: object
      properties:
        name:
          type: string
        sortOrder:
          type: integer
      required:
        - name
    UpdateSafariWorksheetRequest:
      type: object
      properties:
        name:
          type: string
        sortOrder:
          type: integer
        active:
          type: boolean
...