# lib/api-zod/src/generated/types/adminSection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IQmeetEQ Workshop Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminSection {
  id: string;
  title: string;
  code: string;
  codeActive: boolean;
  cohort?: string;
  tier: number;
}