# lib/api-zod/src/generated/types/createSafariWorksheetRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IQmeetEQ Workshop Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSafariWorksheetRequest {
  name: string;
  sortOrder?: number;
}