# lib/api-zod/src/generated/types/safariWorksheet.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IQmeetEQ Workshop Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface SafariWorksheet {
  id: number;
  name: string;
  sortOrder: number;
  active: boolean;
}